ASoC: update aw882xx driver to v0.1.16
Modified content:
Modify the setting of gain initial value, set it according to temperature and voltage,
and modify the setting of volume when fade in.

issue solved:
When the battery power is high and low-temperature and low-pressure protection is not triggered,
the initial GAIN will be set to -6dB.
After the low-temperature and low-pressure trigger, the sound will become louder.

Change-Id: Ia050aa03225c34d63f19c8183585b0be07f5d883
Signed-off-by: xiewen3 <xiewen3@motorola.com>
Reviewed-on: https://gerrit.mot.com/1739266
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Yunfei Guo <guoyf2@motorola.com>
Reviewed-by: Zhengming Yao <yaozm1@mt.com>
Reviewed-by: Xiangpo Zhao <zhaoxp3@motorola.com>
Submit-Approved: Jira Key